Added conditional tippy
