+++ src/components/v2/values/chartValuesDiff/ChartValuesView.reducer.ts
@@ -16,6 +16,7 @@ export const initState = (
         isDeleteInProgress: false,
         showDeleteAppConfirmationDialog: false,
         showRepoSelector: false,
+        showConnectToChartTippy: false,
         selectedProject: null,
         selectedEnvironment: null,
         selectedVersion: selectedVersionFromParent,
@@ -78,6 +79,8 @@ export const chartValuesReducer = (state: ChartValuesViewState, action: ChartVal
             return { ...state, showDeleteAppConfirmationDialog: action.payload }
         case ChartValuesViewActionTypes.showRepoSelector:
             return { ...state, showRepoSelector: action.payload }
+        case ChartValuesViewActionTypes.showConnectToChartTippy:
+            return { ...state, showConnectToChartTippy: action.payload }
         case ChartValuesViewActionTypes.selectedProject:
             return { ...state, selectedProject: action.payload }
         case ChartValuesViewActionTypes.selectedEnvironment:
