+++ src/css/base.scss
@@ -1152,6 +1152,14 @@ button.anchor {
     &.tippy-shadow {
         box-shadow: 0px 2px 4px 0px rgba(0, 0, 0, 0.25);
     }
+
+    .tippy-arrow {
+        color: var(--tippy-card-black-bg);
+
+        &::before {
+            scale: 1.4;
+        }
+    }
 }
 
 .dc__devtron-breadcrumb {
